feat(ast-evaluation): Number is now an Integer, also re-structured the code and added mod and div operators

+++ src/SExpr.hs
@@ -7,7 +7,7 @@ module SExpr (
 import Data.List (intercalate)
 
 data SymbolicExpression
-    = Number Float
+    = Number Integer
     | Symbol String
     | List [SymbolicExpression]
